Updated event hub for #1284 as well, so it will work
