+++ src/MassTransit.WebJobs.EventHubsIntegration/EventHubSendTransport.cs
@@ -13,8 +13,6 @@
 namespace MassTransit.WebJobs.EventHubsIntegration
 {
     using System;
-    using System.Collections.Generic;
-    using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
     using Azure.ServiceBus.Core.Contexts;
